+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Bug fixes:
 * Updated `Kernel.Complex()` to handle the `exception: false` parameter.
 * Fixed `Kernel#dup` to return self for `Complex` and `Rational` objects.
 * Updated `Kernel.Float()` to handle the `exception: false` parameter.
+* Fixed error when `SystemCallError` message contained non-ASCII characters.
 
 Compatibility: